Moving pose stack construction logic to separate PoseStackBuilder class
