+++ tmol/tests/pack/rotamer/dunbrack/test_dunbrack_chi_sampler.py
@@ -9,7 +9,7 @@ from tmol.utility.tensor.common_operations import exclusive_cumsum1d
 
 from tmol.chemical.restypes import RefinedResidueType, ResidueTypeSet
 from tmol.pose.packed_block_types import PackedBlockTypes
-from tmol.pose.pose_stack import PoseStack
+from tmol.pose.pose_stack_builder import PoseStackBuilder
 
 # from tmol.score.coordinates import CartesianAtomicCoordinateProvider
 # from tmol.score.dunbrack.score_graph import DunbrackScoreGraph
@@ -819,9 +819,13 @@ def test_package_samples_for_output(default_database, ubq_res, torch_device):
         for res in ubq_res
     ]
 
-    p1 = PoseStack.one_structure_from_polymeric_residues(ubq_res[5:11], torch_device)
-    p2 = PoseStack.one_structure_from_polymeric_residues(ubq_res[:7], torch_device)
-    poses = PoseStack.from_poses([p1, p2], torch_device)
+    p1 = PoseStackBuilder.one_structure_from_polymeric_residues(
+        ubq_res[5:11], torch_device
+    )
+    p2 = PoseStackBuilder.one_structure_from_polymeric_residues(
+        ubq_res[:7], torch_device
+    )
+    poses = PoseStackBuilder.from_poses([p1, p2], torch_device)
     pbt = poses.packed_block_types
 
     palette = PackerPalette(rts)
@@ -905,9 +909,13 @@ def test_package_samples_for_output(default_database, ubq_res, torch_device):
 
 def test_chi_sampler_smoke(ubq_res, default_database, default_restype_set):
     torch_device = torch.device("cpu")
-    p1 = PoseStack.one_structure_from_polymeric_residues(ubq_res[:5], torch_device)
-    p2 = PoseStack.one_structure_from_polymeric_residues(ubq_res[:7], torch_device)
-    poses = PoseStack.from_poses([p1, p2], torch_device)
+    p1 = PoseStackBuilder.one_structure_from_polymeric_residues(
+        ubq_res[:5], torch_device
+    )
+    p2 = PoseStackBuilder.one_structure_from_polymeric_residues(
+        ubq_res[:7], torch_device
+    )
+    poses = PoseStackBuilder.from_poses([p1, p2], torch_device)
     palette = PackerPalette(default_restype_set)
     task = PackerTask(poses, palette)
     task.restrict_to_repacking()
@@ -930,8 +938,10 @@ def test_chi_sampler_build_lots_of_rotamers(
     # torch_device = torch.device("cpu")
     n_poses = 10
     # print([res.residue_type.name for res in ubq_res[:10]])
-    p = PoseStack.one_structure_from_polymeric_residues(ubq_res[:10], torch_device)
-    poses = PoseStack.from_poses([p] * n_poses, torch_device)
+    p = PoseStackBuilder.one_structure_from_polymeric_residues(
+        ubq_res[:10], torch_device
+    )
+    poses = PoseStackBuilder.from_poses([p] * n_poses, torch_device)
     palette = PackerPalette(default_restype_set)
     task = PackerTask(poses, palette)
     task.restrict_to_repacking()
