+++ tmol/tests/pack/sim_anneal/test_sim_anneal.py
@@ -5,7 +5,7 @@ import numpy
 from tmol.utility.tensor.common_operations import stretch
 
 from tmol.chemical.restypes import ResidueTypeSet
-from tmol.pose.pose_stack import PoseStack
+from tmol.pose.pose_stack_builder import PoseStackBuilder
 
 # to dump pdbs
 # from tmol.system.packed import PackedResidueSystem
@@ -51,8 +51,10 @@ def test_random_rotamer_module(ubq_res, default_database, torch_device):
         for res in ubq_res
     ]
 
-    p = PoseStack.one_structure_from_polymeric_residues(ubq_res[:n_res], torch_device)
-    poses = PoseStack.from_poses([p] * n_poses, torch_device)
+    p = PoseStackBuilder.one_structure_from_polymeric_residues(
+        ubq_res[:n_res], torch_device
+    )
+    poses = PoseStackBuilder.from_poses([p] * n_poses, torch_device)
 
     # contexts = poses.coords.clone() ??
 
@@ -179,8 +181,10 @@ def test_mc_accept_reject_module_smoke(ubq_res, default_database, torch_device):
         for res in ubq_res
     ]
 
-    p = PoseStack.one_structure_from_polymeric_residues(ubq_res[:n_res], torch_device)
-    poses = PoseStack.from_poses([p] * n_poses, torch_device)
+    p = PoseStackBuilder.one_structure_from_polymeric_residues(
+        ubq_res[:n_res], torch_device
+    )
+    poses = PoseStackBuilder.from_poses([p] * n_poses, torch_device)
 
     arange_n_poses = torch.arange(n_poses, dtype=torch.int64, device=torch_device)
 
@@ -261,10 +265,10 @@ def test_accept_final_smoke(
 
     max_n_blocks = 10
     n_poses = 3
-    p = PoseStack.one_structure_from_polymeric_residues(
+    p = PoseStackBuilder.one_structure_from_polymeric_residues(
         rts_ubq_res[:max_n_blocks], torch_device
     )
-    poses = PoseStack.from_poses([p] * n_poses, torch_device)
+    poses = PoseStackBuilder.from_poses([p] * n_poses, torch_device)
 
     palette = PackerPalette(fresh_default_restype_set)
     task = PackerTask(poses, palette)
